+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.3.0 (in development)
 -----------------------
 - Table cell values that begin with a combining character will now have a space
   prepended
+- Added a `padding` option
 
 v0.2.0 (2018-06-16)
 -------------------